Move storageBase to a storage folder where we can keep storage classes
