+++ src/tests/storage/mockStorageBase.ts
@@ -1,6 +1,6 @@
 import {ResponsePackage} from "../../scripts/responsePackage";
 
-import {StorageBase, TimeStampedData} from "../../scripts/extensions/storageBase";
+import {StorageBase, TimeStampedData} from "../../scripts/storage/storageBase";
 
 import {StubSessionLogger} from "../../scripts/logging/stubSessionLogger";
