+++ src/tests/tests.html
@@ -54,10 +54,6 @@
 	<script type="text/javascript" src="extensions/extensionBase_tests.js"></script>
 	<script type="text/javascript" src="extensions/injectHelper_tests.js"></script>
 	<script type="text/javascript" src="extensions/tooltipHelper_tests.js"></script>
-	<script type="text/javascript" src="extensions/storageBase_tests.js"></script>
-
-	<script type="text/javascript" src="versioning/version_tests.js"></script>
-	<script type="text/javascript" src="versioning/changeLogHelper_tests.js"></script>
 
 	<script type="text/javascript" src="localization/localization.js"></script>
 	<script type="text/javascript" src="localization/rtl.js"></script>
@@ -69,6 +65,11 @@
 	<script type="text/javascript" src="logging/loggerDecorator_tests.js"></script>
 	<script type="text/javascript" src="logging/logHelpers_tests.js"></script>
 	<script type="text/javascript" src="logging/sessionLogger_tests.js"></script>
+
+	<script type="text/javascript" src="storage/storageBase_tests.js"></script>
+
+	<script type="text/javascript" src="versioning/version_tests.js"></script>
+	<script type="text/javascript" src="versioning/changeLogHelper_tests.js"></script>
 
 	<script type="text/javascript" src="constants_tests.js"></script>
 	<script type="text/javascript" src="experiments_tests.js"></script>
